+++ web/typescript/index.ts
@@ -2,18 +2,21 @@ import { ComponentMeta, ComponentRegistry } from '@inductiveautomation/perspecti
 import { DateTimeRangePicker, DateTimeRangePickerMeta } from './components/DateTimeRangePicker';
 import { Calendar, CalendarMeta } from './components/Calendar';
 import { ResourceTimeline, ResourceTimelineMeta } from './components/timeline/ResourceTimeline';
+import { DataGrid, DataGridMeta } from './components/grid/DataGrid';
 
 import './scss/main.scss';
 import './scss/calendar.scss';
 import './scss/timeline.scss';
+import './scss/grid.scss';
 
-export { DateTimeRangePicker, Calendar, ResourceTimeline };
+export { DateTimeRangePicker, Calendar, ResourceTimeline, DataGrid };
 
 // Register every component this module provides with the Perspective client registry.
 const components: Array<ComponentMeta> = [
     new DateTimeRangePickerMeta(),
     new CalendarMeta(),
-    new ResourceTimelineMeta()
+    new ResourceTimelineMeta(),
+    new DataGridMeta()
 ];
 
 components.forEach((c: ComponentMeta) => ComponentRegistry.register(c));